+++ dtool/src/prc/configVariableList.h
@@ -39,14 +39,14 @@ PUBLISHED:
                             int flags = 0);
   INLINE ~ConfigVariableList();
 
-  INLINE int get_num_values() const;
-  INLINE string get_string_value(int n) const;
+  INLINE size_t get_num_values() const;
+  INLINE string get_string_value(size_t n) const;
 
-  INLINE int get_num_unique_values() const;
-  INLINE string get_unique_value(int n) const;
+  INLINE size_t get_num_unique_values() const;
+  INLINE string get_unique_value(size_t n) const;
 
-  INLINE int size() const;
-  INLINE string operator [] (int n) const;
+  INLINE size_t size() const;
+  INLINE string operator [] (size_t n) const;
 
   void output(ostream &out) const;
   void write(ostream &out) const;
