+++ panda/src/nativenet/socket_portable.h
@@ -279,16 +279,16 @@ inline int DO_CONNECT(const SOCKET a, const sockaddr_in *b)
 }
 inline int DO_SOCKET_READ(const SOCKET a, char * buf, const int size)
 {
-    return recv(a, buf, size, 0);
+    return (int)recv(a, buf, (size_t)size, 0);
 }
 inline int DO_SOCKET_WRITE(const SOCKET a, const char * buff, const int len)
 {
-    return send(a, buff, len, 0);
+    return (int)send(a, buff, (size_t)len, 0);
 }
 ///////////////////////////////////////////////
 inline int DO_SOCKET_WRITE_TO(const SOCKET a, const char * buffer, const int buf_len, const sockaddr_in * addr)
 {
-    return sendto(a, buffer, buf_len, 0, reinterpret_cast<const struct ::sockaddr *>(addr), sizeof(sockaddr));
+    return (int)sendto(a, buffer, (size_t)buf_len, 0, reinterpret_cast<const struct ::sockaddr *>(addr), sizeof(sockaddr));
 }
 inline SOCKET DO_NEWUDP()
 {
@@ -316,7 +316,7 @@ inline int DO_ACCEPT(SOCKET sck, sockaddr_in * adr)
 inline int DO_RECV_FROM(SOCKET sck, char * data, int len, sockaddr_in * addr)
 {
     socklen_t plen = sizeof(sockaddr);
-    return recvfrom(sck, data, len, 0, (sockaddr *)addr, &plen);
+    return (int)recvfrom(sck, data, (size_t)len, 0, (sockaddr *)addr, &plen);
 }
 
 
@@ -338,7 +338,7 @@ inline int GETERROR()
 
 inline int SOCKIOCTL(const SOCKET s, const long flags, void * val)
 {
-    return ioctl(s, flags, val);
+    return ioctl(s, (unsigned long)flags, val);
 }
 
 inline bool do_shutdown_send(SOCKET s)
